Fixed a problem that caused accidental overwriting of the prototype element.
When directly editing the protoype element, the "save as new" button was activated. However, when inserting new elements into a new collection, this did not happen. Instead they were inserted into the prototype. (LcElement.inc#L531)

Now when working with a prototype elemnt (X1), the create new is performed. This should fix the problem of accidentally overwriting the template.