Fixed a problem with the prelab MC status.
It appears that when we reopened the lab for a team after this team did not start but others already were closed, the labsystem managed to create wrong MC answers after right ones. This should not be possible.
If you update the status now and let the MC update as well, the system is recalculating all prelab statistics.

Added colored info on which MCs were answered wrongly when in seeMe=all mode.

Splitted the update status in the way that you have to select "check prelab" when you want it to be chacked. Should make the update status for the cases where you do not need that much faster.